+++ packages/react-native-sdk/src/components/Participant/index.ts
@@ -1,0 +1,2 @@
+export * from './LocalParticipantView';
+export * from './ParticipantView';