+++ packages/react-native-sdk/src/components/participants/index.ts
@@ -1,2 +1,0 @@
-export * from './LocalParticipantView';
-export * from './Participant';