+++ sample-apps/react/react-video-demo/src/utils/constants.ts
@@ -1,0 +1,3 @@
+export const logoURI = `${
+  import.meta.env.BASE_URL
+}images/icons/stream-logo.svg`;